Make branch picker an activity
This will allow it to be more easily extended and reused.

(cherry picked from commit 490f077bbe3a8d54579ac0f30fe99e3b00fd4d1c)
