+++ src/main/AndroidManifest.xml
@@ -40,6 +40,14 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".activities.RepoDetailActivity" />
         </activity>
+	<activity
+            android:name=".activities.BranchChooserActivity"
+            android:configChanges="orientation|keyboardHidden"
+            android:parentActivityName=".activities.RepoDetailActivity" >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".activities.RepoDetailActivity" />
+        </activity>
         <activity
             android:name=".activities.CommitDiffActivity"
             android:parentActivityName=".activities.RepoDetailActivity" >
